refactor(frontend): further qoL
